+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 [1.1.0] 08.04.2026
 
+- [FAM-6] Wallets in dashboard
 - [FAM-5] Dashboard